stacks/finbot/lambda/bot-req-handler/index.py
# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
# SPDX-License-Identifier: MIT-0

import boto3
import os
import json
import logging
import uuid

logger = logging.getLogger()
logger.setLevel(logging.INFO)

dynamodb_client = boto3.client("dynamodb")


def handler(event, context):
    state_machine_arn = os.environ.get("STATE_MACHINE_ARN")
    # TODO: add code to invoke statemachine.
    return {
        "statusCode": 200,
        "headers": {"Content-Type": "application/json"},
        "body": json.dumps({"message": "TODO call state_machine_arn"}),
    }
